apps/auth/requirements.txt
Flask
influxdb
bcrypt
pyjwt